+++ lib/datasets/cofw.py
@@ -52,8 +52,8 @@ class COFW(data.Dataset):
             self.images = self.mat['IsT']
             self.pts = self.mat['phisT']
 
-        self.mean = np.array([0.485, 0.456, 0.406])
-        self.std = np.array([0.229, 0.224, 0.225])
+        self.mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        self.std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
 
     def __len__(self):
         return len(self.images)
