+++ lib/datasets/wflw.py
@@ -44,8 +44,8 @@ class WFLW(data.Dataset):
         # load annotations
         self.landmarks_frame = pd.read_csv(self.csv_file)
 
-        self.mean = np.array([0.485, 0.456, 0.406])
-        self.std = np.array([0.229, 0.224, 0.225])
+        self.mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        self.std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
 
     def __len__(self):
         return len(self.landmarks_frame)
@@ -66,7 +66,7 @@ class WFLW(data.Dataset):
 
         scale *= 1.25
         nparts = pts.shape[0]
-        img = np.array(Image.open(image_path).convert('RGB'))
+        img = np.array(Image.open(image_path).convert('RGB'), dtype=np.float32)
 
         # transform !
         # img = self.transform(img)
